requirements.txt
python-telegram-bot[webhooks]==21.5
openai==1.51.0
python-dotenv==1.0.1
gunicorn==21.2.0
aiohttp==3.9.1
sqlalchemy==2.0.23
pandas==2.1.4
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==10.1.0



